+++ cookie_consent/util.py
@@ -1,8 +1,7 @@
-import datetime
 import logging
 from collections.abc import Callable, Collection
 
-from django.http import HttpRequest
+from django.http import HttpRequest, HttpResponseBase
 
 from .cache import all_cookie_groups, get_cookie, get_cookie_group
 from .conf import settings
@@ -42,7 +41,7 @@ def _contains_invalid_characters(*inputs: str) -> bool:
     return False
 
 
-def dict_to_cookie_str(dic) -> str:
+def dict_to_cookie_str(dic: dict[str, str]) -> str:
     """
     Serialize a dictionary of cookie-group metadata to a string.
 
@@ -62,24 +61,26 @@ def dict_to_cookie_str(dic) -> str:
     return "|".join(_gen_pairs())
 
 
-def get_cookie_dict_from_request(request):
-    cookie_str = request.COOKIES.get(settings.COOKIE_CONSENT_NAME)
+def get_cookie_dict_from_request(request: HttpRequest) -> dict[str, str]:
+    cookie_str = request.COOKIES.get(settings.COOKIE_CONSENT_NAME, "")
     return parse_cookie_str(cookie_str)
 
 
-def set_cookie_dict_to_response(response, dic):
+def set_cookie_dict_to_response(
+    response: HttpResponseBase, dic: dict[str, str]
+) -> None:
     response.set_cookie(
         settings.COOKIE_CONSENT_NAME,
         dict_to_cookie_str(dic),
         max_age=settings.COOKIE_CONSENT_MAX_AGE,
         domain=settings.COOKIE_CONSENT_DOMAIN,
-        secure=settings.COOKIE_CONSENT_SECURE or None,
-        httponly=settings.COOKIE_CONSENT_HTTPONLY or None,
+        secure=settings.COOKIE_CONSENT_SECURE,
+        httponly=settings.COOKIE_CONSENT_HTTPONLY,
         samesite=settings.COOKIE_CONSENT_SAMESITE,
     )
 
 
-def get_cookie_value_from_request(request, varname, cookie=None):
+def get_cookie_value_from_request(request: HttpRequest, varname: str, cookie: str = ""):
     """
     Returns if cookie group or its specific cookie has been accepted.
 
@@ -121,7 +122,7 @@ def get_cookie_groups(varname: str = "") -> Collection[CookieGroup]:
     return [g for k, g in all_cookie_groups().items() if k in keys]
 
 
-def are_all_cookies_accepted(request):
+def are_all_cookies_accepted(request: HttpRequest) -> bool:
     """
     Returns if all cookies are accepted.
     """
@@ -133,7 +134,7 @@ def are_all_cookies_accepted(request):
     )
 
 
-def _get_cookie_groups_by_state(request, state: bool | None):
+def _get_cookie_groups_by_state(request, state: bool | None) -> Collection[CookieGroup]:
     return [
         cookie_group
         for cookie_group in get_cookie_groups()
@@ -141,62 +142,32 @@ def _get_cookie_groups_by_state(request, state: bool | None):
     ]
 
 
-def get_not_accepted_or_declined_cookie_groups(request):
+def get_not_accepted_or_declined_cookie_groups(
+    request: HttpRequest,
+) -> Collection[CookieGroup]:
     """
     Returns all cookie groups that are neither accepted or declined.
     """
     return _get_cookie_groups_by_state(request, state=None)
 
 
-def get_accepted_cookie_groups(request):
+def get_accepted_cookie_groups(request: HttpRequest) -> Collection[CookieGroup]:
     """
     Returns all cookie groups that are accepted.
     """
     return _get_cookie_groups_by_state(request, state=True)
 
 
-def get_declined_cookie_groups(request):
+def get_declined_cookie_groups(request: HttpRequest) -> Collection[CookieGroup]:
     """
     Returns all cookie groups that are declined.
     """
     return _get_cookie_groups_by_state(request, state=False)
 
 
-def is_cookie_consent_enabled(request: HttpRequest):
+def is_cookie_consent_enabled(request: HttpRequest) -> bool:
     """
     Returns if django-cookie-consent is enabled for given request.
     """
     enabled: bool | Callable[[HttpRequest], bool] = settings.COOKIE_CONSENT_ENABLED
     return enabled(request) if callable(enabled) else enabled
-
-
-# Deprecated
-def get_cookie_string(cookie_dic):
-    """
-    Returns cookie in format suitable for use in javascript.
-    """
-    expires = datetime.datetime.now() + datetime.timedelta(
-        seconds=settings.COOKIE_CONSENT_MAX_AGE
-    )
-    cookie_str = "{}={}; expires={}; path=/".format(
-        settings.COOKIE_CONSENT_NAME,
-        dict_to_cookie_str(cookie_dic),
-        expires.strftime("%a, %d %b %Y %H:%M:%S GMT"),
-    )
-    return cookie_str
-
-
-def get_accepted_cookies(request):
-    """
-    Returns all accepted cookies.
-    """
-    cookie_dic = get_cookie_dict_from_request(request)
-    accepted_cookies = []
-    for cookie_group in all_cookie_groups().values():
-        version = cookie_dic.get(cookie_group.varname, None)
-        if not version or version == settings.COOKIE_CONSENT_DECLINE:
-            continue
-        for cookie in cookie_group.cookie_set.all():
-            if version >= cookie.get_version():
-                accepted_cookies.append(cookie)  # noqa: PERF401
-    return accepted_cookies